+++ nshm_model_graphql_api/schema/nshm_model_schema.py
@@ -10,17 +10,58 @@ from graphene import relay
 log = logging.getLogger(__name__)
 
 
+class SourceLogicTreeBranch(graphene.ObjectType):
+    model_version = graphene.String()
+    weight = graphene.Float()
+
+
+class SourceBranchSet(graphene.ObjectType):
+    model_version = graphene.String()
+    short_name = graphene.String()
+    long_name = graphene.String()
+    branches = graphene.List(SourceLogicTreeBranch)
+
+    def resolve_branches(root, info, **kwargs):
+        log.info(f"resolve_branches root: {root} kwargs: {kwargs}")
+        slt = nm.get_model_version(root.model_version).source_logic_tree
+        for bs in slt.branch_sets:
+            if bs.short_name == root.short_name:
+                for ltb in bs.branches:
+                    yield SourceLogicTreeBranch(root.model_version, ltb.weight)
+
+
+class SourceLogicTree(graphene.ObjectType):
+    model_version = graphene.String()
+    branch_sets = graphene.List(SourceBranchSet)
+
+    @staticmethod
+    def resolve_branch_sets(root, info, **kwargs):
+        log.info(f"resolve_branch_sets root: {root} kwargs: {kwargs}")
+        slt = nm.get_model_version(root.model_version).source_logic_tree
+        for bs in slt.branch_sets:
+            yield SourceBranchSet(root.model_version, bs.short_name, bs.long_name)
+
+
 class NshmModel(graphene.ObjectType):
+    """graphene custom Node representind an entire model."""
 
     class Meta:
         interfaces = (relay.Node,)
 
     version = graphene.String()
     title = graphene.String()
+    source_logic_tree = graphene.Field(SourceLogicTree)
 
     def resolve_id(self, info):
         """resolver for the relay.Node Interface."""
         return self.version
+
+    @staticmethod
+    def resolve_source_logic_tree(root, info, **kwargs):
+        log.info(f"resolve_source_logic_tree root: {root} kwargs: {kwargs}")
+        return SourceLogicTree(
+            model_version=root.version
+        )  # , branch_sets=get_branch_sets(slt))
 
     @classmethod
     def get_node(cls, info, version: str):